First version

+++ src/app/api/categories/route.ts
@@ -1,43 +1,52 @@
 import { NextResponse, NextRequest } from "next/server";
 import Category from "@/models/category";
 import connectMongoDB from "@/libs/mongodb";
+import { utapi } from "uploadthing/server";
 
 export async function GET() {
   await connectMongoDB();
   const categories = await Category.find();
   return NextResponse.json(categories);
 }
-export async function POST(request) {
+export async function POST(request: NextRequest) {
   await connectMongoDB();
   const formData = await request.formData();
-  let body = Object.fromEntries(formData);
+  const name = formData.get("name");
+
+  const files = formData.getAll("file");
+  const uploadedFiles = await utapi.uploadFiles(files);
+  const fileData = uploadedFiles[0].data;
   try {
-    const category = await Category.create({ ...body });
-    return NextResponse.json(category);
+    if (fileData) {
+      const category = await Category.create({ name, image: fileData });
+      return NextResponse.json(category);
+    }
   } catch (error) {
     console.error("Error creating category:", error);
   }
 }
 
-export async function DELETE(req) {
+export async function DELETE(request: NextRequest) {
   await connectMongoDB();
-  const id = req.nextUrl.searchParams.get("id");
+  const id = request.nextUrl.searchParams.get("id");
+  const key = request.nextUrl.searchParams.get("key");
 
   try {
     // Check if the ID is valid
-    if (!id) {
-      throw new Error("Category ID is missing.");
+    if (!id || !key) {
+      throw new Error("Category ID or key is missing.");
     }
 
     const deletedCategory = await Category.findByIdAndDelete(id);
+    await utapi.deleteFiles(key);
 
     if (!deletedCategory) {
       throw new Error("Category not found.");
     }
 
     return NextResponse.json({ message: "Category was deleted" });
-  } catch (error) {
+  } catch (error: any) {
     console.error("Error deleting category:", error);
-    return NextResponse.json({ error: error.message }, 400);
+    return NextResponse.json({ error: error.message });
   }
 }
